requirements.txt
pydantic>=2.0
python-dotenv
httpx
aiohttp
fastapi
uvicorn
rich
py-clob-client
